// cinepath-web/src/app/types/request-types.ts
export interface Keyword {
  id?: number
  name?: string
}

export interface ReleaseDate {
  certification?: string
  iso_639_1?: string
  release_date?: string
  type?: number
  note?: string
}

export interface Video {
  id?: string
  iso_639_1?: string
  iso_3166_1?: string
  key?: string
  name?: string
  official?: boolean
  published_at?: string
  site?: string
  size?: 360 | 480 | 720 | 1080
  type?: 'Trailer' | 'Teaser' | 'Clip' | 'Featurette' | 'Behind the Scenes' | 'Bloopers'
}
// ...
export interface SimpleEpisode {
  air_date?: string
  episode_number?: number
  id?: number
  name?: string
  order?: number
  overview?: string
  production_code?: string
  rating?: number
  season_number?: number
  show_id?: number
  still_path?: string
  vote_average?: number
  vote_count?: number
}

export interface Network {
  name?: string
  id?: number
  logo_path?: string
  origin_country?: string
}

export interface SimpleSeason {
  air_date?: string
  episode_count?: number
  id?: number
  name?: string
  overview?: string
  poster_path?: string
  season_number?: number
}

export interface SimplePerson {
  id?: number
  credit_id?: string
  name?: string
  gender?: number
  profile_path?: string
}

export interface Cast {
  adult?: boolean
  cast_id?: number
  character?: string
  credit_id?: string
  gender?: number | null
  id?: number
  known_for_department?: string
  name?: string
  order?: number
  original_name?: string
  popularity?: number
  profile_path?: string | null
}

export interface Crew {
  adult?: boolean
  credit_id?: string
  department?: string
  gender?: number | null
  id?: number
  known_for_department?: string
  job?: string
  name?: string
  original_name?: string
  popularity?: number
  profile_path?: string | null
}
// ...
export interface PaginatedResponse extends Response {
  page?: number
  total_results?: number
  total_pages?: number
}
// ...
export interface CreditsResponse extends Response {
  id?: number
  cast?: Array<Cast>
  crew?: Array<Crew>
}
// ...
export interface VideosResponse extends Response {
  id?: number
  results?: Array<Video>
}
// ...
export interface ShowResponse extends Response {
  backdrop_path?: string | null
  created_by?: Array<SimplePerson>
  episode_run_time?: number[]
  first_air_date?: string
  genres?: Array<Genre>
  homepage?: string
  id?: number
  in_production?: boolean
  languages?: string[]
  last_air_date?: string
  last_episode_to_air?: SimpleEpisode
  name?: string
  next_episode_to_air?: null
  networks?: Array<Network>
  number_of_episodes?: number
  number_of_seasons?: number
  origin_country?: string[]
  original_language?: string
  original_name?: string
  overview?: string
  popularity?: number
  poster_path?: string | null
  production_companies?: Array<ProductionCompany>
  production_countries?: Array<ProductionCountry>
  seasons?: Array<SimpleSeason>
  spoken_languages?: Array<SpokenLanguage>
  status?: string
  tagline?: string
  type?: string
  vote_average?: number
  vote_count?: number
}
// ...
export interface TvKeywordsResponse extends Response {
  id?: number
  results?: Array<Keyword>
}

export interface TvResultsResponse extends PaginatedResponse {
  results?: Array<TvResult>
}

export interface TvReviewsResponse extends PaginatedResponse {
  results?: Array<Review>
}
// ...
export interface Review {
  id?: string
  author?: string
  content?: string
  iso_639_1?: string
  media_id?: number
  media_title?: string
  media_type?: string
  url?: string
}